refactor(cli): thin group.ts — extract dispatch orchestration to core (S2, #367)
Move runGroupDispatch + helpers (isCaptainAlive, waitForWarmup,
resolveCurrentProject) from cli/commands/group.ts into a new
core library module (packages/core/src/group-dispatch.ts).

The CLI command handler is now parse → bootCaptain closure → dispatchToSibling → format.
CLI-edge concern (shelling out to 'squadrant launch') is injected via bootCaptain callback
so core stays free of exec calls and respects the DAG (shared < core < cli).

Unit tests moved to packages/core/src/__tests__/group-dispatch.test.ts,
mocking ../protocol.js directly instead of @squadrant/core.
CLI test file retains command-registration assertion only.

Build: green. Tests: 1555/1555 (144 files, up from 143).

Part of #367 — more slices remain.
