+++ packages/core/src/index.ts
@@ -24,3 +24,4 @@ export * from "./crew-lifecycle.js";
 export * from "./telegram/index.js";
 export * from "./crew-routing.js";
 export * from "./restart-daemon.js";
+export * from "./group-dispatch.js";
